Move Google Test to deps folder
