+++ deps/googletest/include/gtest/internal/gtest-linked_ptr.h
